requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-JWT-Extended==4.2.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
multitasking==0.0.9
numpy==1.21.0
pandas==1.2.5
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.19
urllib3==1.26.6
Werkzeug==2.0.1
yfinance==0.1.59
